Move deprecated methods into dedicated files
