+++ .github/workflows/llvm-cov.yml
@@ -12,5 +12,5 @@ jobs:
     steps:
       - uses: actions/checkout@v2
       - run: swift test -c release -Xswiftc -enable-testing --enable-code-coverage
-      - run: xcrun llvm-cov export -summary-only -ignore-filename-regex ZIPFoundationTests .build/release/ZIPFoundationPackageTests.xctest/Contents/MacOS/ZIPFoundationPackageTests -instr-profile .build/release/codecov/default.profdata > .build/coverage.json
+      - run: xcrun llvm-cov export -summary-only -ignore-filename-regex 'ZIPFoundationTests|.*Deprecated.*$' .build/release/ZIPFoundationPackageTests.xctest/Contents/MacOS/ZIPFoundationPackageTests -instr-profile .build/release/codecov/default.profdata > .build/coverage.json
       - run: (cat .build/coverage.json|jq '.data[0]["totals"]["lines"]["percent"]' | grep -Eq  "100") && { exit 0; } || { echo 'Please make sure that the test suite covers all framework code paths.'; exit 1; }
