+++ Sources/ZIPFoundation/Data+Compression.swift
@@ -37,7 +37,7 @@ public typealias Consumer = (_ data: Data) throws -> Void
 /// - Throws: Can throw to indicate errors in the data source.
 public typealias Provider = (_ position: Int64, _ size: Int) throws -> Data
 
-extension Data {
+public extension Data {
     enum CompressionError: Error {
         case invalidStream
         case corruptedData
@@ -47,7 +47,7 @@ extension Data {
     ///
     /// - Parameter checksum: The starting seed.
     /// - Returns: The checksum calculated from the bytes of the receiver and the starting seed.
-    public func crc32(checksum: CRC32) -> CRC32 {
+    func crc32(checksum: CRC32) -> CRC32 {
         #if canImport(zlib)
         return withUnsafeBytes { bufferPointer in
             let length = UInt32(count)
@@ -65,21 +65,13 @@ extension Data {
     ///   - provider: A closure that accepts a position and a chunk size. Returns a `Data` chunk.
     ///   - consumer: A closure that processes the result of the compress operation.
     /// - Returns: The checksum of the processed content.
-    public static func compress(size: Int64, bufferSize: Int, provider: Provider, consumer: Consumer) throws -> CRC32 {
+    static func compress(size: Int64, bufferSize: Int, provider: Provider, consumer: Consumer) throws -> CRC32 {
         #if os(macOS) || os(iOS) || os(watchOS) || os(tvOS)
         return try self.process(operation: COMPRESSION_STREAM_ENCODE, size: size, bufferSize: bufferSize,
                                 provider: provider, consumer: consumer)
         #else
         return try self.encode(size: size, bufferSize: bufferSize, provider: provider, consumer: consumer)
         #endif
-    }
-
-    @available(*, deprecated, message: "Please use `Int64` for `size` and provider `position`.")
-    public static func compress(size: Int, bufferSize: Int,
-                                provider: (_ position: Int, _ size: Int) throws -> Data,
-                                consumer: Consumer) throws -> CRC32 {
-        let newProvider: Provider = { try provider(Int($0), $1) }
-        return try self.compress(size: Int64(size), bufferSize: bufferSize, provider: newProvider, consumer: consumer)
     }
 
     /// Decompress the output of `provider` and pass it to `consumer`.
@@ -90,23 +82,14 @@ extension Data {
     ///   - provider: A closure that accepts a position and a chunk size. Returns a `Data` chunk.
     ///   - consumer: A closure that processes the result of the decompress operation.
     /// - Returns: The checksum of the processed content.
-    public static func decompress(size: Int64, bufferSize: Int, skipCRC32: Bool,
-                                  provider: Provider, consumer: Consumer) throws -> CRC32 {
+    static func decompress(size: Int64, bufferSize: Int, skipCRC32: Bool,
+                           provider: Provider, consumer: Consumer) throws -> CRC32 {
         #if os(macOS) || os(iOS) || os(watchOS) || os(tvOS)
         return try self.process(operation: COMPRESSION_STREAM_DECODE, size: size, bufferSize: bufferSize,
                                 skipCRC32: skipCRC32, provider: provider, consumer: consumer)
         #else
         return try self.decode(bufferSize: bufferSize, skipCRC32: skipCRC32, provider: provider, consumer: consumer)
         #endif
-    }
-
-    @available(*, deprecated, message: "Please use `Int64` for `size` and provider `position`.")
-    public static func decompress(size: Int, bufferSize: Int, skipCRC32: Bool,
-                                  provider: (_ position: Int, _ size: Int) throws -> Data,
-                                  consumer: Consumer) throws -> CRC32 {
-        let newProvider: Provider = { try provider(Int($0), $1) }
-        return try self.decompress(size: Int64(size), bufferSize: bufferSize, skipCRC32: skipCRC32,
-                                   provider: newProvider, consumer: consumer)
     }
 }
